// generated/src/app/api-models/member-elig-history.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class MemberEligHistory {
  seqEligHist: number;
  seqMembId: number;
  seqSubsId: number;
  subscriberId: string;
  personNumber: string;
  effectiveDate: Date;
  termDate: Date;
  termReason: string;
  relationshipCode: string;
  seqGroupId: number;
  planCode: string;
  lineOfBusiness: string;
  riderCode1: string;
  riderCode2: string;
  riderCode3: string;
  riderCode4: string;
  riderCode5: string;
  riderCode6: string;
  riderCode7: string;
  riderCode8: string;
  medicareStatusFlg: string;
  otherStatusFlag: string;
  hireDate: Date;
  eligStatus: string;
  premOverrideStep: number;
  premOverrideAmt: number;
  premOverrideCode: string;
  seqProvId: number;
  ipaId: string;
  panelId: string;
  seqProv2Id: number;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  userDefined1: string;
  salary: number;
  pecEndDate: Date;
  reasonCode: string;
  pecWaived: string;
  billEffectiveFromDate: Date;
  billedThruDate: Date;
  paidThruDate: Date;
  subscDept: string;
  subscLocation: string;
  useEftFlg: string;
  benefitStartDate: Date;
  seqEnrollmentRule: number;
  mcareRiskAccretionDate: Date;
  mcareRiskDeletionDate: Date;
  mcareRiskRefusedDate: Date;
  comments: string;
  userDefined2: string;
  userDefined3: string;
  rateType: string;
  pcpaaOccurred: string;
  privacyOn: string;
  pcpChangeReason: string;
  siteCode: string;
  seqSiteAddressId: number;
  entMcarePartAFlg: string;
  entMcarePartBFlg: string;
  entMcareWelfareFlg: string;
  entMcarePipdcgFactor: number;
  medicarePartAEffectDate: Date;
  medicarePartBEnrollDate: Date;
  dateOfDeath: Date;
  facName: string;
  facAddressLine1: string;
  facAddressLine2: string;
  facCity: string;
  facState: string;
  facCounty: string;
  facZipCode: string;
  facCountry: string;
  facPhoneNumber: string;
  facAdmissionDate: Date;
  facDischargeDate: Date;
  facMmCertification: string;
  medRiskUserDefined1: string;
  medRiskUserDefined2: string;
  medRiskUserDate1: Date;
  medRiskUserDate2: Date;
  pbpAcknowledgeIndicator: string;
  mcUserDefined1: string;
  mcUserDefined2: string;
  mcUserDate1: Date;
  mcUserDate2: Date;
  pbpEffectiveDate: Date;
  geocode: string;
  pbp: string;
  facSeqProvId: number;
  csIndicator: string;
  mcarePartABTermDate: Date;
  riderCode9: string;
  riderCode10: string;
  riderCode11: string;
  riderCode12: string;
  riderCode13: string;
  riderCode14: string;
  riderCode15: string;
  riderCode16: string;
  riderCode17: string;
  riderCode18: string;
  riderCode19: string;
  riderCode20: string;
}